updated partially for book inventory

+++ ObserverPattern/src/com/fdmgroup/Observer/Observer.java
@@ -1,7 +1,11 @@
 package com.fdmgroup.Observer;
 
 
-
+/**
+ * 
+ * Interface Observer: Updates Book Stocks in Inventory
+ *
+ */
 public interface Observer {
-	public void update(double ibmPrice, double aaplPrice, double googPrice);
+	public void update(int ibmPrice, int aaplPrice, int googPrice);
 }
